web/requirements.txt
Flask==2.1.3
Flask-SQLAlchemy==2.5.1
psycopg2==2.9.3
PyJWT==2.4.0
Flask-Migrate==3.1.0
Flask-Cors==3.0.10
SQLAlchemy==1.4
flask-restx==0.5.1
redis==4.1.1
rq==1.10.1
psycopg2-binary==2.9.5
celery==5.2.3
Flask-WTF==1.0.0
flower==1.0.0

